Clean null round (#402)
* clean up

* log

* more logs

* more logs

* debug logs
