+++ fendermint/vm/interpreter/src/chain.rs
@@ -266,6 +266,12 @@ where
                         .top_down_msgs_from(prev_height + 1, p.height as u64, &finality.block_hash)
                         .await
                         .context("failed to fetch top down messages")?;
+                    tracing::debug!(
+                        "total topdown msgs: {} from {} to {}",
+                        msgs.len(),
+                        prev_height + 1,
+                        p.height
+                    );
 
                     let ret = topdown::execute_topdown_msgs(&self.gateway_caller, &mut state, msgs)
                         .await
